Rename maven module
