+++ pom.xml
@@ -31,7 +31,7 @@
 		<cloudfoundry-client-lib.version>1.1.3</cloudfoundry-client-lib.version>
 	</properties>
 	<modules>
-		<module>spring-cloud-dataflow-module-registry</module>
+		<module>spring-cloud-dataflow-artifact-registry</module>
 		<module>spring-cloud-dataflow-module-deployers</module>
 		<module>spring-cloud-dataflow-core</module>
 		<module>spring-cloud-dataflow-rest-resource</module>
