+++ spring-cloud-dataflow-admin/pom.xml
@@ -36,7 +36,7 @@
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-dataflow-module-registry</artifactId>
+			<artifactId>spring-cloud-dataflow-artifact-registry</artifactId>
 			<version>1.0.0.BUILD-SNAPSHOT</version>
 		</dependency>
 		<dependency>
