+++ spring-cloud-dataflow-admin/src/test/java/org/springframework/cloud/dataflow/admin/configuration/TestDependencies.java
@@ -19,8 +19,8 @@ package org.springframework.cloud.dataflow.admin.configuration;
 import static org.mockito.Mockito.mock;
 
 import org.springframework.cloud.dataflow.module.deployer.ModuleDeployer;
-import org.springframework.cloud.dataflow.module.registry.InMemoryArtifactRegistry;
-import org.springframework.cloud.dataflow.module.registry.ArtifactRegistry;
+import org.springframework.cloud.dataflow.artifact.registry.InMemoryArtifactRegistry;
+import org.springframework.cloud.dataflow.artifact.registry.ArtifactRegistry;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.web.config.EnableSpringDataWebSupport;
