+++ spring-cloud-dataflow-admin/src/test/java/org/springframework/cloud/dataflow/admin/controller/TaskControllerTests.java
@@ -44,7 +44,7 @@ import org.springframework.cloud.dataflow.core.ModuleDeploymentRequest;
 import org.springframework.cloud.dataflow.core.TaskDefinition;
 import org.springframework.cloud.dataflow.module.deployer.ModuleDeployer;
 import org.springframework.cloud.dataflow.module.deployer.lattice.TaskModuleDeployer;
-import org.springframework.cloud.dataflow.module.registry.InMemoryArtifactRegistry;
+import org.springframework.cloud.dataflow.artifact.registry.InMemoryArtifactRegistry;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.web.WebAppConfiguration;
