+++ spring-cloud-dataflow-artifact-registry/src/main/java/org/springframework/cloud/dataflow/artifact/registry/ArtifactRegistry.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.dataflow.module.registry;
+package org.springframework.cloud.dataflow.artifact.registry;
 
 
 import java.util.List;
