+++ spring-cloud-dataflow-completion/pom.xml
@@ -17,7 +17,7 @@
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-dataflow-module-registry</artifactId>
+			<artifactId>spring-cloud-dataflow-artifact-registry</artifactId>
 			<version>1.0.0.BUILD-SNAPSHOT</version>
 		</dependency>
 		<dependency>
