+++ spring-cloud-dataflow-completion/src/main/java/org/springframework/cloud/dataflow/completion/ChannelNameYieldsModulesRecoveryStrategy.java
@@ -21,8 +21,8 @@ import static org.springframework.cloud.dataflow.core.ArtifactType.*;
 import java.util.List;
 
 import org.springframework.cloud.dataflow.core.dsl.CheckPointedParseException;
-import org.springframework.cloud.dataflow.module.registry.ArtifactRegistration;
-import org.springframework.cloud.dataflow.module.registry.ArtifactRegistry;
+import org.springframework.cloud.dataflow.artifact.registry.ArtifactRegistration;
+import org.springframework.cloud.dataflow.artifact.registry.ArtifactRegistry;
 
 /**
  * Proposes module names when the user has typed a named channel redirection.
