+++ spring-cloud-dataflow-completion/src/main/java/org/springframework/cloud/dataflow/completion/ModulesAfterPipeRecoveryStrategy.java
@@ -22,8 +22,8 @@ import java.util.List;
 
 import org.springframework.cloud.dataflow.core.StreamDefinition;
 import org.springframework.cloud.dataflow.core.dsl.CheckPointedParseException;
-import org.springframework.cloud.dataflow.module.registry.ArtifactRegistration;
-import org.springframework.cloud.dataflow.module.registry.ArtifactRegistry;
+import org.springframework.cloud.dataflow.artifact.registry.ArtifactRegistration;
+import org.springframework.cloud.dataflow.artifact.registry.ArtifactRegistry;
 
 /**
  * Provides completions for the case where the user has entered a pipe
