+++ spring-cloud-dataflow-docs/src/main/asciidoc/spring-cloud-dataflow-overview.adoc
@@ -68,7 +68,7 @@ modules in a linear pipeline from a *source* to a *sink*, optionally including *
 in between. The domain also includes the concept of a **task**, which may be any process that does
 not run indefinitely, including link:https://github.com/spring-projects/spring-batch[Spring Batch] jobs.
 
-The link:https://github.com/spring-cloud/spring-cloud-dataflow/tree/master/spring-cloud-dataflow-module-registry[Module Registry]
+The link:https://github.com/spring-cloud/spring-cloud-dataflow/tree/master/spring-cloud-dataflow-artifact-registry[Artifact Registry]
 maintains the set of available modules, and their mappings to Maven coordinates.
 
 The link:https://github.com/spring-cloud/spring-cloud-dataflow/tree/master/spring-cloud-dataflow-module-deployers/spring-cloud-dataflow-module-deployer-spi[Module Deployer SPI] provides the abstraction layer for deploying the modules of a given stream across a variety of runtime environments, including:
